refactor(comment): move from core to desk
