+++ frappe/desk/doctype/comment/comment.json
@@ -101,7 +101,7 @@
  "links": [],
  "modified": "2023-07-24 13:52:55.651462",
  "modified_by": "Administrator",
- "module": "Core",
+ "module": "Desk",
  "name": "Comment",
  "owner": "Administrator",
  "permissions": [
